Add auto aspect ratio detection.  To use, select an anamorphic or non-anamorphic mode (there's no way to detect *that* automatically!).  Hit "9" or select "Current Image" from the advanced aspect ratio menu.  The aspect ratio will adjust to whatever is currently on the screen.  Note that if you do this with a very dark image, e.g. an outer-space scene, the algorithm may get confused.
Two new ini file settings in the ASPECT section control this feature.

LuminanceThreshold - How black a pixel has to be for the code to consider it black.  Values are 0-255.  Default is 15.

IgnoreNonBlackPixels - Number of non-black pixels to ignore in letterbox areas.  Default is 20% of the image width, which will allow the code to ignore station logos.  If you're never watching material with logos, set this to 1 to improve reliability of the detection code on dark scenes.

This code doesn't continuously detect aspect ratio (at least, not yet) -- you have to manually tell it when to analyze the image.