Automatic aspect ratio detection. When activated (using the "A" key or the "Autodetect" item in the advanced aspect ratio submenu) this code computes the aspect ratio every frame. If it sees that the image needs to be zoomed out, it does so immediately to avoid masking program material. Zooming in is more conservative to prevent aspect ratio thrashing when the edges of a scene are dark; the code has to detect the exact same aspect ratio continuously for a while before it'll zoom in.
This also includes some fixes to the noncontinuous aspect ratio detection submitted previously. I wasn't masking the luminance value correctly, and it turned out to not work so well to just look for the black border at the top of the image.

Two new ASPECT ini file settings:

AutoDetectAspect - if 1, autodetection is activated.  (Toggleable from the UI as described above.) Default is 0.

ZoomInFrameCount - Number of consecutive frames whose aspect ratios have to match before the code will zoom in to the new ratio. Default is 120.