Improve aspect ratio autodetection. A new ini file setting, AspectConsistencyTime (default value 30) is the number of seconds of aspect ratio history the code keeps. Now the code will only switch to a new ratio if:
- No narrower ratio has been detected in the last N seconds, OR
- It has used a ratio at least as wide in the last N seconds.

(Note the distinction between "detected" and "used" here; "used" means that the code actually switched to a ratio.)

In addition, if no narrower ratio has been detected in the last N seconds, the requirement for a precise ratio match for the last ZoomInFrameCount frames is lifted. In practical terms, this allows aspect ratio lockon during the "Star Wars" scrolling text intros, for example, even though the black spaces between lines of text cause the apparent aspect ratio to shift around a bit. The default value of ZoomInFrameCount has been cut in half since the other checks keep the code from zooming in all the time.

This change causes the code to take a lot longer the first time it locks onto an aspect ratio that's wider than a previous one, but it can switch *back* to that ratio very quickly after temporarily dropping to a wider ratio, e.g. for a commercial break. If you don't fast-forward through commercial breaks you'll want to increase AspectConsistencyTime to exceed the length of the breaks.