Still more refinements to auto aspect ratio detection. The code is now much more strict about only switching quickly to ratios it's seen before (it used to be happy to switch to a new ratio if it'd seen anything wider recently).
A new config setting, AspectHistoryTime, controls the length of dTV's memory of old aspect ratios for purposes of quickly switching back to a previously computed ratio. The default is 300 seconds. A ratio has to remain in place for more than 1 second to be entered into the ratio history. The AspectConsistencyTime default has been lowered from 30 to 15 seconds, and its sole purpose now is to specify the amount of time that has to pass without a narrower ratio before a new wide ratio will be used.

There is also a fudge factor for ratio equality comparisons. It's not configurable via ini file for now. Differences of up to 3 pixels' height are considered to be the same ratio. It's used both for deciding to switch to a new ratio (the code won't switch to a new ratio that's very close to the current one) and checking to see if the current ratio has occurred in the past.