Support for both deinterlace algorithms, with preliminary autodetection to switch between them as needed.
Details: Rename the new deinterlace function to DeinterlaceFieldWeave() and the old one to DeinterlaceFieldBob(). Make both functions use the same parameters. Get rid of the loop unrolling in both functions to make them easier to modify, and change the verbatim scanline copying in both functions to use memcpyMMX() which does the write-lots-of-quadwords-at-once magic that prompted the loop unrolling. Add menu items to allow either mode to be selected. Add a shift-spacebar accelerator to cycle backwards through modes (handy for switching back and forth between the two video modes to compare them). Add code to UpdateNTSCPulldownMode() to switch from video-bob to video-weave if the image stays relatively still for a while.