Added a introduction that will serve as a "starter" for anyone willing to change a version of the FAQ ;
It contains, and kept in cvs repository, all updates made to the *MASTER ENGLISH VERSION* off the FAQ.

Every translated version should be upated according to, AND ONLY FROM, this log and the english FAQ.

I repeat ; any update should be done first on the MASTER version which is the english one.

The "last update" comment is added to any FAQ version on the site as soon as it's modified and committed into cvs.
Use this info to check what are the changes that you still haven't reported on the version you maintain.

Please add your name to your updates.

Any questions regarding the FAQ should go to me, Yvonus, at yvonus@removethis_bigfoot.com

Also added last minute changes ;
4/6 again .. Yvonus
. Renamed part "Can I use more than one video card" => ".. more than one video input"
. Added in this part something about cards with multiples connectors (1 or 2 composite, 1 Svhs)
. In the "OS" part, changed the link for the linux port "contact us" linking to "#HELP" anchor
. Removed "Depiste a very powerful .." and include into "picture jumps + dropped frames"