What is the Crashlog.txt file, and where do I find it ?
If one of the developers asks you for a crashlog.txt file you should find this file in the DScaler install directory. This file contains useful information that can be used to find bugs more easily. If you can't find one when you've been asked for one don't worry just inform the developer that one was not created. Also if you have this file but are not observing problems then don't worry.

Where can I download DScaler ?

On DScaler's site, in the download page

History of DScaler

The whole thing started in july 2000. See Cliff Watson's post regarding the home theater computer history in "Who are the pioneers of htpc ?". Ali, one of the first French htpc addicts, was promoting a piece of software called BorgTV (aka bTV) that allowed him to deinterlace analog sources with a level of quality unknown at this time. Thanks to threads like "Not so bad analog video deinterlacing with htpc", people got interested in the subject. Among them was John Adcock.

He released a plugin for bTV that was doing realtime 2:2 pulldown for PAL movies that showed that deinterlacing of such sources was feasable in software (see "John Adock's great PALMovie plugin").

The lack of NTSC support in bTV made the implementation of the more complex 3:2 pulldown algorithm lead John to write his own application based on Multidec. Its name, based on its inspiration application ; dTV (see "First Version of NTSC/PAL TV deinterlacing program").

This proved that complex deinterlacing techniques could be done in realtime and without additionnal dedicated hardware. Since this time, more and more users (and coders) have joined the dTV community and are using it on their htpc rather than using external line doublers.

In may 2001, dTV got renamed to DScaler to avoid any misunderstanding with the growing Digital TeleVision issue.

See Authors for the main contributors and Thanks for a list of people who have helped make DScaler a success.